refactor: 관리자 탭 컴포넌트로 분리
